fix dlib crashing darknet if loaded first
